flask_app/pip_requirements.txt
Flask>=0.10.1
Flask-Mail
Flask-Script
Flask-Migrate
Flask-SQLAlchemy
gevent
Logbook>=0.6.0
PyYAML
requests
urlobject
uwsgi
weber-utils>=1.2.0
